Move: referral state dropdown to new component ReferralStateDropdown
