+++ src/app/referrals/ReferralCard.tsx
@@ -1,12 +1,9 @@
 import { Button } from '@/components/Button';
-import { getConsultByID } from '@/data/consult/getConsult';
-import { patchReferrals } from '@/data/referrals/patchReferral';
+import ReferralStateDropdown from '@/components/referrals/ReferralStateDropdown';
 import { Patient } from '@/types/Patient';
 import { Referral } from '@/types/Referral';
 import Image from 'next/image';
 import Link from 'next/link';
-import { ChangeEvent, useEffect, useState } from 'react';
-import toast from 'react-hot-toast';
 
 export default function ReferralCard({
   ref,
@@ -17,42 +14,6 @@ export default function ReferralCard({
   patient: Patient;
   date: Date;
 }) {
-  const referralState = [
-    'None',
-    'New',
-    'Seen',
-    'Outgoing',
-    'Completed',
-    'CompletedSuccess',
-    'CompletedFailure',
-  ];
-
-  const [referralStatus, setReferralStatus] = useState<string>('');
-
-  useEffect(() => {
-    const fetchConsults = async () => {
-      getConsultByID(ref.consult.toString())
-        .then(() => {
-          setReferralStatus(ref.referral_state);
-        })
-        .catch(e => console.log(e));
-    };
-    fetchConsults();
-  }, []);
-
-  function dropdownChanged(e: ChangeEvent<HTMLSelectElement>) {
-    const patchReferral = async () => {
-      const payload = {
-        ...ref,
-        referral_state: e.target.value,
-      };
-      patchReferrals(payload, ref.id.toString());
-    };
-    patchReferral()
-      .then(() => toast.success('Updated successfully!'))
-      .catch(() => toast.error('Failed to update'));
-  }
-
   return (
     <tr>
       <td>
@@ -88,26 +49,7 @@ export default function ReferralCard({
       </td>
       <td>
         <div className="grid items-center justify-center p-2">
-          <div>
-            <select
-              name="status"
-              id="status"
-              onChange={e => dropdownChanged(e)}
-              className="w-full rounded border-2 p-1"
-            >
-              {referralState.map(status =>
-                status == referralStatus ? (
-                  <option key={status} value={status} selected>
-                    {status}
-                  </option>
-                ) : (
-                  <option key={status} value={status}>
-                    {status}
-                  </option>
-                )
-              )}
-            </select>
-          </div>
+          <ReferralStateDropdown ref={ref} />
         </div>
       </td>
     </tr>
